+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_deployment/online_deployment.py
@@ -254,27 +254,16 @@ class OnlineDeployment(Deployment):
             self.instance_type = other.instance_type or self.instance_type
 
     @classmethod
-    def load(
-        cls,
-        path: Union[PathLike, str] = None,
-        params_override: list = None,
-        **kwargs,
-    ) -> "OnlineDeployment":
-        params_override = params_override or []
-        data = load_yaml(path)
-        return OnlineDeployment.load_from_dict(data=data, path=path, params_override=params_override)
-
-    @classmethod
-    def load_from_dict(
+    def _load(
         cls,
         data: dict,
-        path: Union[PathLike, str] = None,
+        yaml_path: Union[PathLike, str] = None,
         params_override: list = None,
         **kwargs,
     ) -> "OnlineDeployment":
         params_override = params_override or []
         context = {
-            BASE_PATH_CONTEXT_KEY: Path(path).parent if path else Path.cwd(),
+            BASE_PATH_CONTEXT_KEY: Path(yaml_path).parent if yaml_path else Path.cwd(),
             PARAMS_OVERRIDE_KEY: params_override,
         }
 
@@ -667,7 +656,7 @@ class ManagedOnlineDeployment(OnlineDeployment):
         if self.scale_settings:
             if not isinstance(self.scale_settings, DefaultScaleSettings):
                 msg = "ManagedOnlineEndpoint supports DefaultScaleSettings only."
-                raise ValidationtException(
+                raise ValidationException(
                     message=msg,
                     target=ErrorTarget.ONLINE_DEPLOYMENT,
                     no_personal_data_message=msg,