+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_text_analytics_client.py
@@ -58,14 +58,11 @@ from ._models import (
     ExtractKeyPhrasesAction,
     AnalyzeSentimentAction,
     AnalyzeHealthcareEntitiesResult,
-    ExtractSummaryAction,
-    ExtractSummaryResult,
     RecognizeCustomEntitiesAction,
     RecognizeCustomEntitiesResult,
     SingleCategoryClassifyAction,
-    SingleCategoryClassifyResult,
     MultiCategoryClassifyAction,
-    MultiCategoryClassifyResult,
+    ClassifyDocumentResult,
     AnalyzeHealthcareEntitiesAction,
     _AnalyzeActionsType,
 )
@@ -83,10 +80,8 @@ AnalyzeActionsResponse = AnalyzeActionsLROPoller[
                 RecognizePiiEntitiesResult,
                 ExtractKeyPhrasesResult,
                 AnalyzeSentimentResult,
-                ExtractSummaryResult,
                 RecognizeCustomEntitiesResult,
-                SingleCategoryClassifyResult,
-                MultiCategoryClassifyResult,
+                ClassifyDocumentResult,
                 AnalyzeHealthcareEntitiesResult,
                 DocumentError,
             ]
@@ -1072,7 +1067,6 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                 RecognizePiiEntitiesAction,
                 ExtractKeyPhrasesAction,
                 AnalyzeSentimentAction,
-                ExtractSummaryAction,
                 RecognizeCustomEntitiesAction,
                 SingleCategoryClassifyAction,
                 MultiCategoryClassifyAction,
@@ -1089,10 +1083,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                     RecognizePiiEntitiesResult,
                     ExtractKeyPhrasesResult,
                     AnalyzeSentimentResult,
-                    ExtractSummaryResult,
                     RecognizeCustomEntitiesResult,
-                    SingleCategoryClassifyResult,
-                    MultiCategoryClassifyResult,
+                    ClassifyDocumentResult,
                     AnalyzeHealthcareEntitiesResult,
                     DocumentError,
                 ]
@@ -1120,7 +1112,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             The action results will be in the same order of the input actions.
         :type actions:
             list[RecognizeEntitiesAction or RecognizePiiEntitiesAction or ExtractKeyPhrasesAction or
-            RecognizeLinkedEntitiesAction or AnalyzeSentimentAction or ExtractSummaryAction or
+            RecognizeLinkedEntitiesAction or AnalyzeSentimentAction or
             RecognizeCustomEntitiesAction or SingleCategoryClassifyAction or
             MultiCategoryClassifyAction or AnalyzeHealthcareEntitiesAction]
         :keyword str display_name: An optional display name to set for the requested analysis.
@@ -1150,18 +1142,17 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         :rtype:
             ~azure.ai.textanalytics.AnalyzeActionsLROPoller[~azure.core.paging.ItemPaged[
             list[RecognizeEntitiesResult or RecognizeLinkedEntitiesResult or RecognizePiiEntitiesResult or
-            ExtractKeyPhrasesResult or AnalyzeSentimentResult or ExtractSummaryAction or RecognizeCustomEntitiesResult
-            or SingleCategoryClassifyResult or MultiCategoryClassifyResult or AnalyzeHealthcareEntitiesResult or
-            DocumentError]]]
+            ExtractKeyPhrasesResult or AnalyzeSentimentResult or RecognizeCustomEntitiesResult
+            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or DocumentError]]]
         :raises ~azure.core.exceptions.HttpResponseError or TypeError or ValueError:
 
         .. versionadded:: v3.1
             The *begin_analyze_actions* client method.
         .. versionadded:: 2022-04-01-preview
-            The *ExtractSummaryAction*, *RecognizeCustomEntitiesAction*, *SingleCategoryClassifyAction*,
+            The *RecognizeCustomEntitiesAction*, *SingleCategoryClassifyAction*,
             *MultiCategoryClassifyAction*, and *AnalyzeHealthcareEntitiesAction* input options and the
-            corresponding *ExtractSummaryResult*, *RecognizeCustomEntitiesResult*, *SingleCategoryClassifyResult*,
-            *MultiCategoryClassifyResult*, and *AnalyzeHealthcareEntitiesResult* result objects
+            corresponding *RecognizeCustomEntitiesResult*, *ClassifyDocumentResult*,
+            and *AnalyzeHealthcareEntitiesResult* result objects
 
         .. admonition:: Example:
 
@@ -1277,22 +1268,6 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                 sentiment_analysis_tasks=[
                     a for a in generated_tasks
                     if _determine_action_type(a) == _AnalyzeActionsType.ANALYZE_SENTIMENT
-                ],
-                extractive_summarization_tasks=[
-                    a for a in generated_tasks
-                    if _determine_action_type(a) == _AnalyzeActionsType.EXTRACT_SUMMARY
-                ],
-                custom_entity_recognition_tasks=[
-                    a for a in generated_tasks
-                    if _determine_action_type(a) == _AnalyzeActionsType.RECOGNIZE_CUSTOM_ENTITIES
-                ],
-                custom_single_classification_tasks=[
-                    a for a in generated_tasks
-                    if _determine_action_type(a) == _AnalyzeActionsType.SINGLE_CATEGORY_CLASSIFY
-                ],
-                custom_multi_classification_tasks=[
-                    a for a in generated_tasks
-                    if _determine_action_type(a) == _AnalyzeActionsType.MULTI_CATEGORY_CLASSIFY
                 ],
             )
             analyze_body = models.AnalyzeBatchInput(