+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Release History
 
-## 12.13.0b1 (Unreleased)
+## 12.13.0b1 (2022-06-15)
 
 ### Features Added
 - Added support for service version 2021-08-06.
+- Added a new version of client-side encryption for blobs (version 2.0) which utilizes AES-GCM-256 encryption.
+If you are currently using client-side encryption, it is **highly recommended** to switch to a form of server-side
+encryption (Customer-Provided Key, Encryption Scope, etc.) or version 2.0 of client-side encryption. The encryption
+version can be specified on any client constructor via the `encryption_version` keyword (`encryption_version='2.0'`).
 
 ## 12.12.0 (2022-05-09)